requirements.txt
# Core
anthropic>=0.40.0
httpx>=0.27.0

# Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document Processing
unstructured[all-docs]>=0.15.0
python-magic>=0.4.27

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Utils
tqdm>=4.66.0
python-dotenv>=1.0.0
